# src/gen/java/io/swagger/model/Offer.java
/*
 * Swagger Offers
 * This is a demo Offers  server developed by atilaek.   [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).
 *
 * OpenAPI spec version: v1
 * Contact: atilaek@atilek.atilaek
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

package io.swagger.model;

import com.fasterxml.jackson.annotation.JsonProperty;
import io.swagger.v3.oas.annotations.media.Schema;

import javax.validation.constraints.NotNull;
import java.util.Date;
import java.util.Objects;

/**
 * Offer
 */
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.JavaJerseyServerCodegen", date = "2018-09-24T13:09:22.776Z[GMT]")
public class Offer {
    @JsonProperty("id")
    private Long id = null;

    @JsonProperty("name")
    private String name = null;

    @JsonProperty("price")
    private OfferPrice price = null;

    @JsonProperty("description")
    private String description = null;

    @JsonProperty("expirationDate")
    private Date expirationDate = null;

    @JsonProperty("cancelled")
    private Boolean cancelled = false;

    public Offer id(Long id) {
        this.id = id;
        return this;
    }

    /**
     * Get id
     *
     * @return id
     **/
    @JsonProperty("id")
    @Schema(description = "")
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public Offer name(String name) {
        this.name = name;
        return this;
    }

    /**
     * Get name
     *
     * @return name
     **/
    @JsonProperty("name")
    @Schema(example = "Chip - Orange Merhcant", required = true, description = "")
    @NotNull
    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Offer price(OfferPrice price) {
        this.price = price;
        return this;
    }

    /**
     * Get price
     *
     * @return price
     **/
    @JsonProperty("price")
    @Schema(required = true, description = "")
    @NotNull
    public OfferPrice getPrice() {
        return price;
    }

    public void setPrice(OfferPrice price) {
        this.price = price;
    }

    public Offer description(String description) {
        this.description = description;
        return this;
    }

    /**
     * Get description
     *
     * @return description
     **/
    @JsonProperty("description")
    @Schema(example = "Offer for Orange Merhcant per a chip", description = "")
    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Offer expirationDate(Date expirationDate) {
        this.expirationDate = expirationDate;
        return this;
    }

    /**
     * Get expirationDate
     *
     * @return expirationDate
     **/
    @JsonProperty("expirationDate")
    @Schema(description = "")
    public Date getExpirationDate() {
        return expirationDate;
    }

    public void setExpirationDate(Date expirationDate) {
        this.expirationDate = expirationDate;
    }

    /**
     * Checks if the offers expiration date is passed.
     *
     * @return @{@link Boolean}
     */
    public Boolean isOutdated() {
        return getExpirationDate().before(new Date());
    }


    public Offer cancelled(Boolean cancelled) {
        this.cancelled = cancelled;
        return this;
    }

    /**
     * Get cancelled
     *
     * @return cancelled
     **/
    @JsonProperty("cancelled")
    @Schema(description = "")
    public Boolean isCancelled() {
        return cancelled;
    }

    public void setCancelled(Boolean cancelled) {
        this.cancelled = cancelled;
    }


    @Override
    public boolean equals(java.lang.Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        Offer offer = (Offer) o;
        return Objects.equals(this.id, offer.id) &&
                Objects.equals(this.name, offer.name) &&
                Objects.equals(this.price, offer.price) &&
                Objects.equals(this.description, offer.description) &&
                Objects.equals(this.expirationDate, offer.expirationDate) &&
                Objects.equals(this.cancelled, offer.cancelled);
    }

    @Override
    public int hashCode() {
        return Objects.hash(id, name, price, description, expirationDate, cancelled);
    }


    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder();
        sb.append("class Offer {\n");

        sb.append("    id: ").append(toIndentedString(id)).append("\n");
        sb.append("    name: ").append(toIndentedString(name)).append("\n");
        sb.append("    price: ").append(toIndentedString(price)).append("\n");
        sb.append("    description: ").append(toIndentedString(description)).append("\n");
        sb.append("    expirationDate: ").append(toIndentedString(expirationDate)).append("\n");
        sb.append("    cancelled: ").append(toIndentedString(cancelled)).append("\n");
        sb.append("}");
        return sb.toString();
    }

    /**
     * Convert the given object to string with each line indented by 4 spaces
     * (except the first line).
     */
    private String toIndentedString(java.lang.Object o) {
        if (o == null) {
            return "null";
        }
        return o.toString().replace("\n", "\n    ");
    }
}